webhook/requirements.txt
Flask==1.1.1
ffmpeg==1.4
gcloud==0.18.3
google-api-core==1.14.3                   
google-auth==1.8.1                    
google-cloud-speech==1.3.1
google-cloud-storage==1.23.0                     
googleapis-common-protos==1.6.0 
line-bot-sdk==1.15.0
pydub==0.23.1
speechrecognition==3.8.1